requirements.txt
Django==1.7
django-mptt
django-haystack
requests
pyelasticsearch
django-bootstrap-form
django-tastypie
django-toolbelt